Slight improvement of isPointOnTriangle Efficiency
Switched to a strategy of getting the sum of the smallest angles between each of the vectors from the point to the vertexes of the triangle. If it is 360, it is in the triangle, but if it is less, that means the vectors are all in the same direction so it is outside the triangle.